Changed functionality of the file
In early development, it was another way to do what OctaneGGStats.py did. Now, OctaneGG reads a python list that was output from this file. This file now writes a python file called RLDataSet.py that sets up usage for OctaneGG. This program should be run anytime a new set of stats are needed